requirements.txt
numpy>=1.24
pandas>=2.0
torch>=2.0
tensorflow>=2.15
scikit-learn>=1.3
matplotlib>=3.7
plotly>=5.18
streamlit>=1.30
streamlit-folium>=0.18
folium>=0.15
requests>=2.31
pyyaml>=6.0
